pyproject.toml
[project]
name = "zzz-snake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gymnasium>=1.1.1",
    "mss>=10.0.0",
    "opencv-python>=4.11.0.86",
    "pynput>=1.8.1",
    "pyyaml>=6.0.2",
    "stable-baselines3>=2.6.0",
    "tensorboard>=2.19.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
